DSS_gRPC_Server/requirements.txt
backports.functools-lru-cache==1.5
certifi==2019.3.9
chardet==3.0.4
cntk==2.6
cntk-gpu==2.6
cycler==0.10.0
docopt==0.6.2
enum34==1.1.6
future==0.17.1
futures==3.2.0
grpc==0.3.post19
grpcio==1.16.0
grpcio-tools==1.16.0
h5py==2.8.0
idna==2.8
jieba==0.39
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
matplotlib==2.2.3
msgpack-python==0.5.6
numpy==1.15.4
pandas==0.23.4
pathlib2==2.3.0
pipreqs==0.4.9
protobuf==3.6.1
protobuf-tools==0.0.3
pyclips==1.0.7.343
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
requests==2.21.0
scandir==1.9.0
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
urllib3==1.24.1
yarg==0.1.9
